pyproject.toml
[project]
name = "fasthtml_playground"
version = "0.0.1"
description = "FastHTML Playground"

# NOTE: Overwrites .python-version
requires-python = ">=3.12,<3.13"

authors = [
    {name = "İlker SIĞIRCI", email = "sigirci.ilker@gmail.com"},
]
maintainers = [
    {name = "İlker SIĞIRCI", email = "sigirci.ilker@gmail.com"},
]
readme = "README.md"

keywords = [
    "fasthtml",
    "playground"
]

classifiers = [
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "loguru>=0.7.3",
    "python-fasthtml>=0.12.0",
    "tinyredis>=0.0.2",
    "yt-dlp>=2025.1.15",
]

[tool.uv]
default-groups = ["dev", "doc", "test"]


[tool.pytest.ini_options]
minversion = "7.0.0"
addopts = [
    "--import-mode=importlib",
    "--color=yes",
    "--verbose",
#    "--durations=0",
#    "--strict-markers",
#    "--doctest-modules",
    # NOTE: cov is not working with pycharm debugger. Enable it later!
#    "--cov api",
#    "--cov-report=xml",
#    "--cov-report html",
]

testpaths = ["tests"]

# TODO: Make them inherit from yaml file
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)-20s :: %(name)-8s :: %(levelname)-8s :: %(filename)s:%(lineno)d - (%(funcName)s) :: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
# target-version = "py312"
line-length = 88
# src = ["api"]
respect-gitignore = true

# extend-include = []

extend-exclude = [
    "docs",
    "tests",
]

# "path/to/file.py" = ["E402"]
# 'api/__init__.py' = ['F405', 'F403']

# [tool.ruff.flake8-quotes]
# inline-quotes = "single"
# multiline-quotes = "double"
# docstring-quotes = "double"

[tool.ruff.format]
skip-magic-trailing-comma = false
# indent-style = "space"
# preview = true

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "C",  # flake8-comprehensions
    "C9",  # mccabe
    "D",  # flake8-docstrings
    "E",  # pycodestyle errors (default)
    "F",  # pyflakes (default)
    "I",  # isort
    "PD", # pandas-vet
    "PIE", # pie
    "PL",  # pylint
    "PTH", # pathlib
    "Q",  # flake8-quotes
    "RET", # return
    "RUF", # Enable all ruff-specific checks
    "SIM", # simplify
    "S307", # eval
    "T20",  # (disallow print statements) keep debugging statements out of the codebase
    "W",  # pycodestyle warnings
    "ASYNC" # async
]

ignore = [
    "E501",  # Line too long, handled by ruff formatter
    "D107", # "Missing docstring in __init__",
    "F811", # "redefinition of the same function"
]

unfixable = [
    "F401" # Unused imports.
]

#extend-select = []

# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"*.ipynb" = ["F401"]
"api/*" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
combine-as-imports = true
# force-wrap-aliases = true
# known-first-party = ["SomeLibrary"]
# known-third-party = ["SomeLibrary"]

# mypy section
# Read more here: https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml-file
[tool.mypy]
# python_version = "3.8"
files = ["api"]
# exclude = ['^tests/']
strict = true
show_error_codes = true
# ignore_missing_imports = true

# TODO: Check whether this is working
[[tool.mypy.overrides]]
module = "tests.*"
ignore_missing_imports = true
check_untyped_defs = true

[tool.coverage.run]
source = ['api']
branch = true
context = '${CONTEXT}'

[tool.coverage.report]
precision = 2
exclude_lines = [
    'pragma: no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING:',
    '@overload',
]
# show_missing = True

# [tool.coverage.paths]
# source = [
#     'api/',
# ]

# [build-system]
# requires = ["uv>=0.5.0,<1.0.0"]
# build-backend = "uv" # FIXME: Only builds source distribution, not wheel

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "ruff==0.9.2",
    # "scalene~=1.5.21.2",
]
doc = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.42",
    "mkdocstrings[python]>=0.26.2",
    # "mdx-include",
    # "mkdocs-markdownextradata-plugin",
]
test = [
    "pytest-cov>=5.0.0",
    "pytest-rerunfailures>=14.0",
    "pytest-xdist>=3.6.1",
    "pytest>=8.3.3",
    # "pytest-timeout",
]
